requirements.txt
Django>=6.0,<7
djangorestframework>=3.15,<4
Pillow>=11.0
